refactor duplicaed code across sat proof checkers into a common base class
