+++ benchexec/tools/cake_pb_cnf.py
@@ -5,11 +5,10 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import benchexec.result as result
-import benchexec.tools.template
+from benchexec.tools.sat_proof_checker_base import SatProofCheckerBase
 
 
-class Tool(benchexec.tools.template.BaseTool2):
+class Tool(SatProofCheckerBase):
     """
     CakePB is a formally verified checker for pseudo-Boolean unsatisfiability proofs,
     providing machine-checked correctness guarantees for the VeriPB proof format.
@@ -23,20 +22,3 @@ class Tool(benchexec.tools.template.BaseTool2):
 
     def project_url(self):
         return "https://gitlab.com/MIAOresearch/software/cakepb"
-
-    def cmdline(self, executable, options, task, rlimits):
-        return [executable, task.single_input_file, *options]
-
-    def determine_result(self, run):
-        for line in run.output:
-            if line.startswith("s "):
-                verdict = line.strip().split(" ")[1].strip().upper()
-                try:
-                    sat_arg = f"({line.strip().split(' ')[2].strip().upper()})"
-                except IndexError:
-                    sat_arg = ""
-                if verdict == "VERIFIED":
-                    return result.RESULT_TRUE_PROP + sat_arg
-                elif verdict == "ERROR":
-                    return result.RESULT_ERROR
-        return result.RESULT_ERROR
