+++ benchexec/tools/gratgen.py
@@ -5,11 +5,10 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import benchexec.result as result
-import benchexec.tools.template
+from benchexec.tools.sat_proof_checker_base import SatProofCheckerBase
 
 
-class Tool(benchexec.tools.template.BaseTool2):
+class Tool(SatProofCheckerBase):
     """
     GRAT-Gen converts standard DRAT unsatisfiability proofs into the GRAT format,
     enabling efficient verification by the formally verified GRATchk checker.
@@ -23,20 +22,3 @@ class Tool(benchexec.tools.template.BaseTool2):
 
     def project_url(self):
         return "https://www21.in.tum.de/~lammich/grat/"
-
-    def cmdline(self, executable, options, task, rlimits):
-        return [executable, task.single_input_file, *options]
-
-    def determine_result(self, run):
-        for line in run.output:
-            if line.startswith("s "):
-                verdict = line.strip().split(" ")[1].strip().upper()
-                try:
-                    sat_arg = f"({line.strip().split(' ')[2].strip().upper()})"
-                except IndexError:
-                    sat_arg = ""
-                if verdict == "VERIFIED":
-                    return result.RESULT_TRUE_PROP + sat_arg
-                elif verdict == "ERROR":
-                    return result.RESULT_ERROR
-        return result.RESULT_ERROR
